requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
yt-dlp==2024.5.27
spotipy==2.23.0
python-dotenv==1.0.1
